Slight code and file reorganization
