--- src/models/Role.js
@@ -1,6 +1,0 @@
-const Role = {
-    ADMIN: "admin",
-    USER: "user",
-};
-
-module.exports = Role;
