+++ src/routes/index.js
@@ -1,8 +1,10 @@
 const express = require("express");
-const authRouter = require("./auth");
+const AuthenticationRouter = require("./AuthenticationRouter");
+const UserRouter = require("./UserRouter");
 
 const app = express();
 
-app.use("/auth", authRouter);
+app.use("/auth", AuthenticationRouter);
+app.use("/user", UserRouter);
 
 module.exports = app;
